feat : 퀴즈 카테고리 삭제 테스트 코드 추가

+++ cs25-service/src/main/java/com/example/cs25service/domain/quiz/service/QuizCategoryService.java
@@ -71,6 +71,9 @@ public class QuizCategoryService {
 
     @Transactional
     public void deleteQuizCategory(Long quizCategoryId){
+        if (!quizCategoryRepository.existsById(quizCategoryId)) {
+            throw new QuizException(QuizExceptionCode.QUIZ_CATEGORY_NOT_FOUND_ERROR);
+        }
         quizCategoryRepository.deleteById(quizCategoryId);
     }
 }
